Fixed a bug where a connection is reachable only using another connection of the same trip that is not the earliest arrival at the origin. This occurs when there is a faster arrival at the origin using another trip, but the interchange time makes the connection unreachable, whereas staying on the marginally slower connection with the same trip would make it reachable because interchange time would not be applied.
An example of this can be seen between LEI and LIV on 2020-06-12. At 14:50 there should be a change at CRE using the S10735 trip at 16:58, however another service arrives marginaly earlier at CRE but is unable to make the connection to depart at 16:58 because of interchange time. Avoiding the change to get to CRE faster allows the passenger to stay on the S10735 from LEI and make the connection. By tracking the best arrival time for each trip at each stop we can decide whether it would be better to stay on that trip in order to get the improved arrival times at later stops.